update transaction test program
Removed hardcoded variables specific to PostgreSQL.  Fixed various
locations where incorrect datatypes were being used, which truncated
longs to ints.  Fixed hardcoded locations for EGen's flat_in inputs.

+++ code/src/transactions/CustomerPositionDB.cpp
@@ -56,7 +56,7 @@ void CCustomerPositionDB::DoCustomerPositionFrame1(
 	int i = 0;
 	for  (p = vAux.begin(); p != vAux.end(); ++p)
 	{
-		pOut->acct_id[i] = atoi( (*p).c_str() );
+		pOut->acct_id[i] = atol( (*p).c_str() );
 		++i;
 	}
 	vAux.clear();
